refactor: :truck: Grouping component in dirs
